move to quarterdeck
